fix(Chip): Move all `Chip` css to one file
