+++ packages/react/src/components/Chip/Chip.module.css
@@ -1,3 +1,136 @@
 .chip {
   display: flex;
 }
+
+.chipButton {
+  --fdsc-chip-height: var(--fds-sizing-7);
+  --fdsc-chip-padding: var(--fds-spacing-3);
+  --fdsc-chip-background: var(--fds-semantic-surface-action-first-subtle);
+  --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
+  --fdsc-chip-border: var(--fds-semantic-border-action-first-subtle);
+  --fdsc-chip-border-radius: var(--fds-border_radius-full);
+
+  background: var(--fdsc-chip-background);
+  padding: 0 var(--fdsc-chip-padding);
+  min-height: var(--fdsc-chip-height);
+  border-radius: var(--fdsc-chip-border-radius);
+  border: 1px solid var(--fdsc-chip-border);
+  color: var(--fdsc-chip-text-color);
+  text-decoration: none;
+  font-family: inherit;
+}
+
+/* Only use hover for non-touch devices to prevent sticky-hovering */
+@media (hover: hover) and (pointer: fine) {
+  .chipButton:hover {
+    --fdsc-chip-background: var(--fds-semantic-surface-action-first-subtle-hover);
+    --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
+    --fdsc-chip-border: var(--fds-semantic-border-action-first-subtle-hover);
+
+    cursor: pointer;
+  }
+
+  .chipButton:is([aria-pressed='true']):hover {
+    --fdsc-chip-background: var(--fds-semantic-surface-action-first-no_fill-active);
+    --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
+  }
+}
+
+.chipButton:is([aria-pressed='true']),
+.chipButton:active {
+  --fdsc-chip-background: var(--fds-semantic-surface-action-first-active);
+  --fdsc-chip-text-color: var(--fds-semantic-text-neutral-on_inverted);
+  --fdsc-chip-border: var(--fds-semantic-surface-action-first-active);
+}
+
+.chipButton .label {
+  color: inherit;
+  line-height: normal;
+  display: flex;
+  align-items: center;
+  flex-direction: row;
+  gap: var(--fds-spacing-2);
+  padding: var(--fds-spacing-1) 0;
+}
+
+.xsmall {
+  --fdsc-chip-height: var(--fds-sizing-6);
+  --fdsc-chip-padding: var(--fds-spacing-2);
+}
+
+.small {
+  --fdsc-chip-height: var(--fds-sizing-7);
+  --fdsc-chip-padding: var(--fds-spacing-3);
+}
+
+.removable {
+  --fdsc-removable-background: var(--fds-semantic-surface-action-first-active);
+  --fdsc-removable-text-color: var(--fds-semantic-text-neutral-on_inverted);
+  --fdsc-removable-chip-size: var(--fds-sizing-7);
+  --fdsc-removable-chip-xmark-color: var(--fds-semantic-text-neutral-on_inverted);
+  --fdsc-removable-chip-xmark-padding_right: var(--fds-spacing-1);
+  --fdsc-removable-chip-xmark-size: var(--fds-sizing-5);
+  --fdsc-removable-chip-xmark-wrapper-width: calc(var(--fdsc-removable-chip-xmark-size) + var(--fdsc-removable-chip-xmark-padding_right));
+
+  color: var(--fdsc-removable-text-color);
+  background: var(--fdsc-removable-background);
+  border: 0;
+  padding-right: 0;
+  min-height: var(--fdsc-removable-chip-size);
+}
+
+.xMark {
+  color: var(--fdsc-removable-chip-xmark-color);
+  display: inline-flex;
+  padding-right: var(--fdsc-removable-chip-xmark-padding_right);
+}
+
+.removable.xsmall {
+  --fdsc-removable-chip-size: var(--fds-sizing-6);
+}
+
+.removable.small {
+  --fdsc-removable-chip-size: var(--fds-sizing-7);
+}
+
+.removable:hover,
+.removable:focus {
+  --fdsc-removable-background: linear-gradient(
+    to left,
+    var(--fds-semantic-surface-danger-default) var(--fdsc-removable-chip-xmark-wrapper-width),
+    var(--fds-semantic-surface-action-first-active) var(--fdsc-removable-chip-xmark-wrapper-width)
+  );
+  --fdsc-removable-chip-xmark-color: var(--fds-semantic-text-neutral-on_inverted);
+}
+
+.xMark .icon {
+  height: var(--fdsc-removable-chip-xmark-size);
+  width: var(--fdsc-removable-chip-xmark-size);
+}
+
+.spacing {
+  padding-left: var(--fds-spacing-1) !important;
+}
+
+.checkmarkIcon {
+  height: var(--fds-sizing-5);
+  width: auto;
+}
+
+.groupContainer {
+  --fdsc-gap: var(--fds-spacing-2);
+
+  display: flex;
+  gap: var(--fdsc-gap);
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+.groupContainer.xsmall {
+  --fdsc-gap: var(--fds-spacing-2);
+}
+
+.groupContainer.small {
+  --fdsc-gap: var(--fds-spacing-3);
+}
