--- packages/react/src/components/Chip/Group/Group.module.css
@@ -1,17 +1,0 @@
-.groupContainer {
-  --fdsc-gap: var(--fds-spacing-2);
-
-  display: flex;
-  gap: var(--fdsc-gap);
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.groupContainer.xsmall {
-  --fdsc-gap: var(--fds-spacing-2);
-}
-
-.groupContainer.small {
-  --fdsc-gap: var(--fds-spacing-3);
-}
