--- packages/react/src/components/Chip/Removable/Removable.module.css
@@ -1,44 +1,0 @@
-.removable {
-  --fdsc-removable-background: var(--fds-semantic-surface-action-first-active);
-  --fdsc-removable-text-color: var(--fds-semantic-text-neutral-on_inverted);
-  --fdsc-removable-chip-size: var(--fds-sizing-7);
-  --fdsc-removable-chip-xmark-color: var(--fds-semantic-text-neutral-on_inverted);
-  --fdsc-removable-chip-xmark-padding_right: var(--fds-spacing-1);
-  --fdsc-removable-chip-xmark-size: var(--fds-sizing-5);
-  --fdsc-removable-chip-xmark-wrapper-width: calc(var(--fdsc-removable-chip-xmark-size) + var(--fdsc-removable-chip-xmark-padding_right));
-
-  color: var(--fdsc-removable-text-color);
-  background: var(--fdsc-removable-background);
-  border: 0;
-  padding-right: 0;
-  min-height: var(--fdsc-removable-chip-size);
-}
-
-.xMark {
-  color: var(--fdsc-removable-chip-xmark-color);
-  display: inline-flex;
-  padding-right: var(--fdsc-removable-chip-xmark-padding_right);
-}
-
-.removable.xsmall {
-  --fdsc-removable-chip-size: var(--fds-sizing-6);
-}
-
-.removable.small {
-  --fdsc-removable-chip-size: var(--fds-sizing-7);
-}
-
-.removable:hover,
-.removable:focus {
-  --fdsc-removable-background: linear-gradient(
-    to left,
-    var(--fds-semantic-surface-danger-default) var(--fdsc-removable-chip-xmark-wrapper-width),
-    var(--fds-semantic-surface-action-first-active) var(--fdsc-removable-chip-xmark-wrapper-width)
-  );
-  --fdsc-removable-chip-xmark-color: var(--fds-semantic-text-neutral-on_inverted);
-}
-
-.xMark .icon {
-  height: var(--fdsc-removable-chip-xmark-size);
-  width: var(--fdsc-removable-chip-xmark-size);
-}
