--- packages/react/src/components/Chip/_ChipBase/ChipBase.module.css
@@ -1,60 +1,0 @@
-.chipButton {
-  --fdsc-chip-height: var(--fds-sizing-7);
-  --fdsc-chip-padding: var(--fds-spacing-3);
-  --fdsc-chip-background: var(--fds-semantic-surface-action-first-subtle);
-  --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
-  --fdsc-chip-border: var(--fds-semantic-border-action-first-subtle);
-  --fdsc-chip-border-radius: var(--fds-border_radius-full);
-
-  background: var(--fdsc-chip-background);
-  padding: 0 var(--fdsc-chip-padding);
-  min-height: var(--fdsc-chip-height);
-  border-radius: var(--fdsc-chip-border-radius);
-  border: 1px solid var(--fdsc-chip-border);
-  color: var(--fdsc-chip-text-color);
-  text-decoration: none;
-  font-family: inherit;
-}
-
-/* Only use hover for non-touch devices to prevent sticky-hovering */
-@media (hover: hover) and (pointer: fine) {
-  .chipButton:hover {
-    --fdsc-chip-background: var(--fds-semantic-surface-action-first-subtle-hover);
-    --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
-    --fdsc-chip-border: var(--fds-semantic-border-action-first-subtle-hover);
-
-    cursor: pointer;
-  }
-
-  .chipButton:is([aria-pressed='true']):hover {
-    --fdsc-chip-background: var(--fds-semantic-surface-action-first-no_fill-active);
-    --fdsc-chip-text-color: var(--fds-semantic-text-neutral-default);
-  }
-}
-
-.chipButton:is([aria-pressed='true']),
-.chipButton:active {
-  --fdsc-chip-background: var(--fds-semantic-surface-action-first-active);
-  --fdsc-chip-text-color: var(--fds-semantic-text-neutral-on_inverted);
-  --fdsc-chip-border: var(--fds-semantic-surface-action-first-active);
-}
-
-.chipButton .label {
-  color: inherit;
-  line-height: normal;
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-  gap: var(--fds-spacing-2);
-  padding: var(--fds-spacing-1) 0;
-}
-
-.xsmall {
-  --fdsc-chip-height: var(--fds-sizing-6);
-  --fdsc-chip-padding: var(--fds-spacing-2);
-}
-
-.small {
-  --fdsc-chip-height: var(--fds-sizing-7);
-  --fdsc-chip-padding: var(--fds-spacing-3);
-}
